src/jobapplier/ai-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
sentence-transformers>=2.2.2
numpy>=1.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
pdfminer.six>=20241005
python-docx>=0.8.11
spacy>=3.7.2
pygls>=1.0.0
transformers>=4.49.0
torch>=2.1.2
